backend/requirements.txt
alembic==1.18.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.31.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
fastapi==0.128.0
google-ai-generativelanguage
google-api-core
google-api-python-client
google-auth
google-auth-httplib2
google-generativeai
google-genai
googleapis-common-protos
greenlet
grpcio
grpcio-status
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
idna==3.11
jiter==0.12.0
kiwipiepy==0.22.2
Mako==1.3.10
MarkupSafe==3.0.3
numpy>=2.2.0,<2.4
openai==2.15.0
pgvector==0.4.2
proto-plus
protobuf
psycopg2-binary==2.9.11
pyasn1==0.6.2
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
pyparsing==3.3.1
python-dotenv==1.2.1
requests==2.32.5
rsa==4.9.1
sniffio==1.3.1
SQLAlchemy==2.0.45
sqlmodel==0.0.31
starlette==0.50.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3>=1.26.5,<1.27
uvicorn==0.40.0
# 인증/보안
python-jose[cryptography]
passlib[bcrypt]>=1.7.4
# 소켓 통신
python-socketio[client]
# S3 Storage
aioboto3==11.3.0
